% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NimbleFunctions.R
\name{dbern_vec}
\alias{dbern_vec}
\title{Internal functions for Nimble: these are the underlying model}
\usage{
dbern_vec(x, p, wt, log = 0)
}
\description{
Internal functions for Nimble: these are the underlying model
}
