% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitMaxEnt.R
\name{FitMaxEnt}
\alias{FitMaxEnt}
\title{Fit Bayesian MaxEnt model with MCMC}
\usage{
FitMaxEnt(maxdat, parallel = FALSE, nchains = 1, nCores = nchains, ...)
}
\arguments{
\item{maxdat}{Data (usually created by SetUpMaxEnt()).}

\item{parallel}{Logical. Should the chains be run in parallel? Defaults to FALSE}

\item{nchains}{Integer. How many chains to run. Defaults to 1}

\item{nCores}{Integer. How many cores to use. Defaults to 1, ignored if parallel=FALSE}

\item{...}{Other arguments passed to runMCMC() (e.g. niter, nburnin)}
}
\value{
A coda object.
}
\description{
Fit Bayesian MaxEnt model with MCMC
}
\examples{
\dontrun{
FitMaxEnt(Data=X, code=MaxNetcode,
adaptInterval=5, nchains=1, nburnin = 5, niter=10, thin=1)
}
}
