% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NimbleFunctions.R
\docType{data}
\name{MaxNetcode}
\alias{MaxNetcode}
\alias{dbern_vec}
\alias{rbern_vec}
\title{Internal Nimble code}
\format{
An object of class \code{{} of length 8.
}
\usage{
MaxNetcode

dbern_vec(x, p, wt, log = 0)

rbern_vec(n, p, wt)
}
\arguments{
\item{...}{Please ignore. I'm just trying to make R happy.}
}
\value{
Functions for Nimble
}
\description{
These should be internal functions, but if they are not exported,
they cannot always be found.
}
\details{
As they are exported they need a help page.
}
\section{Functions}{
\itemize{
\item \code{dbern_vec()}: vectorised weighted Bernoulli density using in the Nimble code

\item \code{rbern_vec()}: vectorised random Bernoulli distribution used by Nimble

}}
\examples{
\dontrun{
NimbleMaxEnt(Data=X, code=MaxNetcode, adaptInterval=1e3,
nchains=1, nburnin = 1e3, niter=6e3, thin=1)
}
}
\keyword{datasets}
