% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NimbleMaxEnt.R
\name{NimbleMaxEnt}
\alias{NimbleMaxEnt}
\title{Run one chain of a Bayesian MaxEnt model}
\usage{
NimbleMaxEnt(Data, code, adaptInterval = 1000, ...)
}
\arguments{
\item{Data}{Data (usually created by SetUpMaxEnt()).}

\item{code}{Nimble code}

\item{adaptInterval}{Integer. How many iterations to adapt for. Defaults to 1000}

\item{...}{Other arguments passed to runMCMC() (e.g. niter, nburnin)}
}
\value{
A coda object.
}
\description{
This should normally be used within \code{FitMaxEnt()}
}
\examples{
\dontrun{
NimbleMaxEnt(Data=X, code=MaxNetcode, adaptInterval=1e3,
nchains=1, nburnin = 1e3, niter=6e3, thin=1)
}
}
