% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetUpMaxEnt.R
\name{SetUpMaxEnt}
\alias{SetUpMaxEnt}
\title{Function to set up data for use in a Bayesian MaxEnt}
\usage{
SetUpMaxEnt(
  p,
  data,
  f = maxnet.formula(p, data, ...),
  regmult = 1,
  lambda = 1,
  regfun = maxnet.default.regularization,
  addsamplestobackground = TRUE,
  ...
)
}
\arguments{
\item{p}{a vector of 1 (for presence) or 0 (for background).}

\item{data}{a matrix or data frame of predictor variables.}

\item{f}{a formula to determine the features to be used. Has a default, so can be ignored}

\item{regmult}{a constant to adjust regularization.}

\item{lambda}{Nimble code}

\item{regfun}{a function to compute regularization constant for each feature. Defaults tov maxnet's default}

\item{addsamplestobackground}{if T, add to the background any presence sample that is not already there.}

\item{...}{Other arguments, passed to \code{maxnet.formula()}}
}
\value{
A list with lists Const, Data, and Inits
}
\description{
This formats the data, including creating the basis functions that MaxNet uses.
The code is partly taken from the \code{maxnet::maxnet()} function.
}
\details{
To change the basis functions (e.g. only include linear terms) use \code{classes=}
(e.g. \code{classes="l"} for linear terms).
}
\examples{
\dontrun{
NimbleMaxEnt(Data=ToNimble, code=MaxNetcode, adaptInterval=1e3,
nchains=1, nburnin = 1e3, niter=6e3, thin=1)
}
}
